ai-service/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-dotenv
httpx
prometheus-client
